Implement zstd decompression for worker results in UI pipeline
