+++ package.json
@@ -72,6 +72,7 @@
     "fflate": "^0.6.9",
     "file-saver": "^2.0.5",
     "filereader-stream": "^2.0.0",
+    "fzstd": "^0.1.1",
     "identity-obj-proxy": "^3.0.0",
     "immer": "^9.0.6",
     "ini": "^2.0.0",
@@ -133,7 +134,8 @@
     "vega-webgl-renderer": "^1.0.0-beta.2",
     "virtualizedtableforantd4": "^1.2.2",
     "webpack-bundle-analyzer": "^4.7.0",
-    "zarrita": "^0.3.2"
+    "zarrita": "^0.3.2",
+    "zstd-codec": "^0.1.5"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.20.2",
