+++ src/utils/work/unpackResult.js
@@ -1,19 +1,25 @@
 import { decompress } from 'fflate';
+import { isZstdCompressed } from './fzstdDecompress';
+
 
 const unpackResult = async (storageResp, taskName = null) => {
-  // SeuratObject can fail to download when loaded into memory
   if (taskName === 'DownloadAnnotSeuratObject' || taskName === 'GetNormalizedExpression') {
     const blob = await storageResp.blob();
     return (blob);
   }
 
   const arrayBuf = await storageResp.arrayBuffer();
-
-  return decompressUint8Array(new Uint8Array(arrayBuf));
+  const uint8 = new Uint8Array(arrayBuf);
+  const decompressed = await decompressUint8Array(uint8);
+  return decompressed;
 };
 
-const decompressUint8Array = async (array) => (
-  new Promise((resolve, reject) => {
+
+const decompressUint8Array = async (array) => {
+  if (isZstdCompressed(array)) {
+    return array; // Pass raw to zstd parser downstream
+  }
+  return new Promise((resolve, reject) => {
     decompress(array, (err, decompressed) => {
       if (err) {
         reject(err);
@@ -21,8 +27,8 @@ const decompressUint8Array = async (array) => (
         resolve(decompressed);
       }
     });
-  })
-);
+  });
+};
 
 export { decompressUint8Array };
 export default unpackResult;
